[TM-1605] add external component to progress and goals site
