+++ src/pages/site/[uuid]/tabs/GoalsAndProgress.tsx
@@ -2,7 +2,6 @@ import { useT } from "@transifex/react";
 import React from "react";
 import { Else, If, Then, When } from "react-if";
 
-import ProgressGoalsDoughnutChart from "@/admin/components/ResourceTabs/MonitoredTab/components/ProgressGoalsDoughnutChart";
 import GoalProgressCard from "@/components/elements/Cards/GoalProgressCard/GoalProgressCard";
 import Text from "@/components/elements/Text/Text";
 import { IconNames } from "@/components/extensive/Icon/Icon";
@@ -12,6 +11,8 @@ import PageRow from "@/components/extensive/PageElements/Row/PageRow";
 import TreeSpeciesTablePD from "@/components/extensive/Tables/TreeSpeciesTablePD";
 import { Framework } from "@/context/framework.provider";
 import { TextVariants } from "@/types/common";
+
+import GoalsAndProgressSiteTab from "../components/GoalsAndProgressSiteTab";
 
 interface GoalsAndProgressTabProps {
   site: any;
@@ -31,40 +32,6 @@ export const LABEL_LEGEND = [
     color: "bg-secondary-600"
   }
 ];
-interface ChartDataItem {
-  cardValues: {
-    label: string;
-    value: number;
-    totalName?: string;
-    totalValue?: number;
-  };
-  chartData: any;
-  graph?: boolean;
-  hectares?: boolean;
-}
-
-type ChartsData = {
-  terrafund: JSX.Element[];
-  ppc: JSX.Element[];
-  hbf: JSX.Element[];
-};
-
-const CharData = (values: ChartDataItem) => {
-  return (
-    <GoalProgressCard
-      label={values.cardValues.label}
-      value={values.cardValues.value}
-      totalValue={values.cardValues.totalValue}
-      hectares={values.hectares}
-      graph={values.graph}
-      classNameLabel="text-neutral-650 uppercase mb-3"
-      labelVariant="text-14"
-      classNameCard="text-center flex flex-col items-center"
-      classNameLabelValue="justify-center"
-      chart={<ProgressGoalsDoughnutChart key={"items"} data={values.chartData} />}
-    />
-  );
-};
 
 const GoalsAndProgressTab = ({ site }: GoalsAndProgressTabProps) => {
   const t = useT();
@@ -140,143 +107,12 @@ const GoalsAndProgressTab = ({ site }: GoalsAndProgressTabProps) => {
       nonTreeCount: "7,500"
     }
   ];
-  const chartDataHectares = {
-    chartData: [
-      { name: t("HECTARES RESTORED"), value: site.total_hectares_restored_sum },
-      site.framework_key !== Framework.PPC
-        ? {
-            name: t("TOTAL HECTARES RESTORED"),
-            value: parseFloat(site.hectares_to_restore_goal)
-          }
-        : {}
-    ],
-    cardValues: {
-      label: t("HECTARES RESTORED"),
-      value: site.total_hectares_restored_sum,
-      totalName: t("TOTAL HECTARES RESTORED"),
-      totalValue: parseFloat(site.hectares_to_restore_goal)
-    }
-  };
-  const chartDataTreesRestored = {
-    chartData: [
-      { name: t("TREES RESTORED"), value: site.trees_restored_count },
-      site.framework_key == Framework.HBF
-        ? {
-            name: t("TOTAL TREES RESTORED"),
-            value: 0
-          }
-        : {}
-    ],
-    cardValues: {
-      label: t("TREES RESTORED"),
-      value: site.trees_restored_count
-    }
-  };
-  const chartDataWorkdays = {
-    chartData: [{ name: t("WORKDAYS CREATED"), value: site.workday_count }],
-    cardValues: {
-      label: t("WORKDAYS CREATED"),
-      value: site.workday_count
-    }
-  };
-  const chartDataSaplings = {
-    chartData: [
-      { name: t("SAPLINGS RESTORED"), value: site.sapling_species_count }
-      // {
-      //   name: t("TOTAL SAPLINGS RESTORED"),
-      //   value: 200
-      // }
-    ],
-    cardValues: {
-      label: t("SAPLINGS RESTORED"),
-      value: site.sapling_species_count
-      // totalName: t("TOTAL SAPLINGS RESTORED"),
-      // totalValue: 200
-    }
-  };
-
-  const chartsDataMapping: ChartsData = {
-    terrafund: [
-      <CharData
-        key={"terrafund-1"}
-        cardValues={chartDataHectares.cardValues}
-        chartData={chartDataHectares}
-        hectares={true}
-      />,
-      <CharData
-        key={"terrafund-2"}
-        cardValues={chartDataTreesRestored.cardValues}
-        chartData={chartDataTreesRestored}
-        graph={false}
-      />
-    ],
-    ppc: [
-      <CharData
-        key={"ppc-1"}
-        cardValues={chartDataHectares.cardValues}
-        chartData={chartDataHectares}
-        graph={false}
-        hectares={true}
-      />,
-      <CharData
-        key={"ppc-2"}
-        cardValues={chartDataTreesRestored.cardValues}
-        chartData={chartDataTreesRestored}
-        graph={false}
-      />,
-      <CharData key={"ppc-3"} cardValues={chartDataWorkdays.cardValues} chartData={chartDataWorkdays} graph={false} />
-    ],
-    hbf: [
-      <CharData key={"hbf-1"} cardValues={chartDataWorkdays.cardValues} chartData={chartDataWorkdays} graph={false} />,
-      <CharData
-        key={"hbf-2"}
-        cardValues={chartDataHectares.cardValues}
-        chartData={chartDataHectares}
-        hectares={true}
-      />,
-      <CharData key={"hbf-3"} cardValues={chartDataSaplings.cardValues} chartData={chartDataSaplings} graph={false} />
-    ]
-  };
 
   return (
     <PageBody>
       <PageRow>
         <PageCard title={t("Progress & Goals")}>
-          <div className="flex w-full flex-wrap items-start justify-between gap-8">
-            {chartsDataMapping[site.framework_key as keyof ChartsData].map((chart, index) => (
-              <React.Fragment key={index}>{chart}</React.Fragment>
-            ))}
-            <GoalProgressCard
-              label={t("Trees restored")}
-              value={site.trees_restored_count}
-              limit={site.trees_grown_goal}
-              hasProgress={false}
-              items={[
-                {
-                  iconName: IconNames.TREE_CIRCLE_PD,
-                  label: t("Trees Planted:"),
-                  variantLabel: "text-14",
-                  classNameLabel: " text-neutral-650 uppercase",
-                  value: site.trees_planted_count
-                },
-                {
-                  iconName: IconNames.LEAF_CIRCLE_PD,
-                  label: t("Seeds Planted:"),
-                  variantLabel: "text-14",
-                  classNameLabel: " text-neutral-650 uppercase",
-                  value: site.seeds_planted_count
-                },
-                {
-                  iconName: IconNames.REFRESH_CIRCLE_PD,
-                  label: t("Trees Regenerating:"),
-                  variantLabel: "text-14",
-                  classNameLabel: " text-neutral-650 uppercase",
-                  value: site.regenerated_trees_count
-                }
-              ]}
-              className="pr-[41px] lg:pr-[150px]"
-            />
-          </div>
+          <GoalsAndProgressSiteTab site={site} />
         </PageCard>
       </PageRow>
